% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umls_api_search_functions.R
\name{get_service_ticket}
\alias{get_service_ticket}
\title{Functions that Make and Parse API Calls}
\usage{
get_service_ticket(store_creds_at = "~")
}
\description{
Functions that Make and Parse API Calls
}
\seealso{
\code{\link[httr]{POST}},\code{\link[httr]{content}},\code{\link[httr]{status_code}}
\code{\link[rvest]{html_nodes}},\code{\link[rvest]{html_text}}
}
