% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umls_api_lookup_functions.R
\name{list_source_subsets}
\alias{list_source_subsets}
\title{List Source Subsets}
\usage{
list_source_subsets(
  vocabulary,
  pageNumber = NULL,
  pageSize = NULL,
  language = "ENG"
)
}
\arguments{
\item{pageNumber}{(optional)  Whole number that specifies which page of results to fetch.}

\item{pageSize}{(optional)  Whole number that specifies the number of results to include per page.}

\item{language}{(optional)  3-letter abbreviation for language}
}
\description{
List Source Subsets
}
\details{
'/subsets/current'path Retrieves information about all subsets from the current release and returns a JSON Object classType of Subset
'/subsets/current/source/SNOMEDCT_US/6011000124106'path Retrieves information for a SNOMED CT subset and returns a JSON Object classType of Subset
'/subsets/current/source/SNOMEDCT_US/6011000124106/members'path Retrieves members of a SNOMED CT subset and returns a JSON Object classType of SourceConceptSubsetMember
'/subsets/current/source/SNOMEDCT_US/6011000124106/member/89361000119103' path Retrieves an individual member of a SNOMED CT subset and returns a JSON Object classType of SourceConceptSubsetMember
}
\seealso{
\code{\link[httr]{GET}},\code{\link[httr]{content}}
}
