% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umls_api_lookup_functions.R
\name{lookup_atoms}
\alias{lookup_atoms}
\title{List Atoms}
\usage{
lookup_atoms(
  AUI,
  sabs = NULL,
  ttys = NULL,
  language = NULL,
  includeObsolete = NULL,
  includeSuppressible = NULL,
  pageNumber = NULL,
  pageSize = NULL
)
}
\arguments{
\item{sabs}{(optional)  Comma-separated list of source vocabularies to include in your search}

\item{ttys}{(optional)  One or more term types}

\item{language}{(optional)  Retrieve only atoms that have a specific language}

\item{includeObsolete}{(optional)  Include content that is obsolete according to the content provider or NLM.}

\item{includeSuppressible}{(optional)  Include content that is suppressible according to NLM Editors.}

\item{pageNumber}{(optional)  Whole number that specifies which page of results to fetch.}

\item{pageSize}{(optional)  Whole number that specifies the number of results to include per page.}
}
\description{
List Atoms
}
\details{
'/content/current/CUI/C0155502/atoms'path Retrieves all atoms for C0155502 and returns a JSON Object classType of Atom
'/content/current/CUI/C0155502/atoms/preferred'path Retrieves the default preferred atom of C0155502 and returns a JSON Object classType of Atom
'/content/current/CUI/C0155502/atoms?language=ENG'path Retrieves all English language atoms for C0155502 and returns a JSON Object classType of Atom
'/content/current/CUI/C0155502/atoms?sabs=SNOMEDCT_US,ICD9CM&ttys=PT'path Retrieve SNOMEDCT_US and ICD9CM preferred terms in C0155502 and returns a JSON Object classType of Atom
'/content/current/source/SNOMEDCT_US/111541001/atoms'path Retrieve atoms belonging to SNOMED CT concept 111541001 and returns a JSON Object classType of Atom
'/content/current/source/SNOMEDCT_US/111541001/atoms/preferred'path Retrieve the default preferred atom belonging to SNOMED CT concept 111541001 and returns a JSON Object classType of Atom
'/content/current/AUI/A8345234'path Retrieve information about AUI A8345234 and returns a JSON Object classType of Atom
'/content/current/AUI/A8345234/ancestors'path Retrieve ancestors of a UMLS atom and returns a JSON Object classType of Atom
'/content/current/AUI/A8345234/descendants'path Retrieve descendants of a UMLS atom and returns a JSON Object classType of Atom
'/content/current/AUI/A8345234/parents'path Retrieve parents of a UMLS atom and returns a JSON Object classType of Atom
'/content/current/AUI/A8345234/children'path Retrieve children of a UMLS atom and returns a JSON Object classType of Atom
}
\seealso{
\code{\link[httr]{GET}},\code{\link[httr]{content}}
}
