% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umls_api_lookup_functions.R
\name{lookup_concept}
\alias{lookup_concept}
\title{Lookup a Concept}
\usage{
lookup_concept(CUI)
}
\description{
Lookup a Concept
}
\details{
'/content/current/CUI/C0009044'path Retrieves CUI and returns a JSON Object classType of Concept
'/content/current/CUI/C0009044/atoms'path Retrieve atoms in a CUI and returns a JSON Object classType of Atom
'/content/current/CUI/C0009044/definitions'path Retrieve CUI definitions and returns a JSON Object classType of Definition
'/content/current/CUI/C0009044/relations'path Retrieve CUI relations and returns a JSON Object classType of ConceptRelation
}
\seealso{
\code{\link[httr]{GET}},\code{\link[httr]{content}}
}
