% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umls_api_lookup_functions.R
\name{lookup_content_view_members}
\alias{lookup_content_view_members}
\title{Lookup Content Views}
\usage{
lookup_content_view_members(CUI, pageNumber = NULL, pageSize = NULL)
}
\arguments{
\item{pageNumber}{(optional)  Whole number that specifies which page of results to fetch.}

\item{pageSize}{(optional)  Whole number that specifies the number of results to include per page.}
}
\description{
Lookup Content Views
}
\details{
'/content-views/current' path retrieves information about all subsets from the current release and returns a JSON Object classType of ContentView
'/content-views/current/CUI/C2711988' path Retrieves information for the SNOMED CT CORE Problem List content view and returns a JSON Object classType of ContentView
'/content-views/current/CUI/C2711988/members' path Retrieves members of the SNOMED CT CORE Problem List content view and returns a JSON Object classType of SourceConceptContentViewMember
'/content-views/current/CUI/C2711988/member/238788004' path Retrieves an individual member of the SNOMED CT CORE Problem List content view and returns a JSON Object classType of SourceConceptContentViewMember
}
\seealso{
\code{\link[httr]{GET}},\code{\link[httr]{content}}
}
