% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umls_api_lookup_functions.R
\name{lookup_crosswalk}
\alias{lookup_crosswalk}
\title{Lookup a Crosswalk}
\usage{
lookup_crosswalk(
  targetSource = NULL,
  includeObsolete = NULL,
  pageNumber = NULL,
  pageSize = NULL
)
}
\arguments{
\item{targetSource}{(optional)  Returns codes from the specified UMLS vocabulary}

\item{includeObsolete}{(optional)  Determines whether to return obsolete codes.}

\item{pageNumber}{(optional)  Whole number that specifies which page of results to fetch.}

\item{pageSize}{(optional)  Whole number that specifies the number of results to include per page.}
}
\description{
Lookup a Crosswalk
}
\details{
'/crosswalk/current/source/HPO/HP:0001947'path Retrieves all codes that share a UMLS CUI with HP:0001947 and returns a JSON Object classType of SourceAtomCluster
'/crosswalk/current/source/HPO/HP:0001947?targetSource=SNOMEDCT_US'path Retrieves all SNOMEDCT_US codes that share a UMLS CUI with HP:0001947 and returns a JSON Object classType of SourceAtomCluster
}
\seealso{
\code{\link[httr]{GET}},\code{\link[httr]{content}}
}
