% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umls_api_lookup_functions.R
\name{lookup_definitions}
\alias{lookup_definitions}
\title{Lookup Definitions}
\usage{
lookup_definitions(sabs = NULL, pageNumber = NULL, pageSize = NULL)
}
\arguments{
\item{sabs}{(optional)  Comma-separated list of source vocabularies to include in your search}

\item{pageNumber}{(optional)  Whole number that specifies which page of results to fetch.}

\item{pageSize}{(optional)  Whole number that specifies the number of results to include per page.}
}
\description{
Lookup Definitions
}
\details{
'/content/current/CUI/C0155502/definitions'path Retrieves definitions of the CUI and returns a JSON Object classType of Definition
}
\seealso{
\code{\link[httr]{GET}},\code{\link[httr]{content}}
}
