% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umls_api_lookup_functions.R
\name{lookup_relations}
\alias{lookup_relations}
\title{Lookup Relations}
\usage{
lookup_relations(pageNumber = NULL, pageSize = NULL)
}
\arguments{
\item{pageNumber}{(optional)  Whole number that specifies which page of results to fetch.}

\item{pageSize}{(optional)  Whole number that specifies the number of results to include per page.}

\item{ticket}{A single-use service ticket is required for each call to the API. See authentication for more information}
}
\description{
Lookup Relations
}
\details{
'/content/current/CUI/C0009044/relations' path Retrieves NLM-asserted relationships of the CUI and returns a JSON Object classType of ConceptRelation
}
\seealso{
\code{\link[httr]{GET}},\code{\link[httr]{content}}
}
