% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umls_api_lookup_functions.R
\name{lookup_semantic_network}
\alias{lookup_semantic_network}
\title{Semantic Network}
\usage{
lookup_semantic_network()
}
\description{
Semantic Network
}
\details{
/semantic-network/current/TUI/T109' path Retrieves TUI and returns a JSON Object classType of SemanticType
}
\seealso{
\code{\link[httr]{GET}},\code{\link[httr]{content}}
}
