% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umls_api_lookup_functions.R
\name{lookup_source_asserted_identifiers}
\alias{lookup_source_asserted_identifiers}
\title{Lookup Source Asserted Identifiers}
\usage{
lookup_source_asserted_identifiers(sourceId, sab)
}
\description{
Lookup Source Asserted Identifiers
}
\details{
'/content/current/source/SNOMEDCT_US/9468002'path Retrieves Source Concept and returns a JSON Object classType of SourceAtomCluster
'/content/current/source/MSH/D015242'path Retrieves Source Descriptor and returns a JSON Object classType of SourceAtomCluster
'/content/current/source/LNC/54112-8'path Retrieves Code and returns a JSON Object classType of SourceAtomCluster
'/content/current/source/SNOMEDCT_US/9468002/atoms'path Retrieve atoms in a source-asserted identifier and returns a JSON Object classType of Atom
'/content/current/source/SNOMEDCT_US/9468002/parents'path Retrieve immediate parents of a source-asserted identifier and returns a JSON Object classType of SourceAtomCluster
'/content/current/source/SNOMEDCT_US/9468002/children'path Retrieve immediate children of source-asserted identifier and returns a JSON Object classType of SourceAtomCluster
'/content/current/source/SNOMEDCT_US/9468002/ancestors'path Retrieve all ancestors of a source-asserted identifier and returns a JSON Object classType of SourceAtomCluster
'/content/current/source/SNOMEDCT_US/9468002/descendants'path Retrieve all descendants of source-asserted identifier and returns a JSON Object classType of SourceAtomCluster
'/content/current/source/SNOMEDCT_US/9468002/attributes'path Retrieves information about source-asserted attributes and returns a JSON Object classType of Attribute
}
\seealso{
\code{\link[httr]{GET}},\code{\link[httr]{content}}
}
