% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umls_api_lookup_functions.R
\name{lookup_source_attributes}
\alias{lookup_source_attributes}
\title{Lookup Source Attributes}
\usage{
lookup_source_attributes(
  sourceId,
  vocabulary,
  pageNumber = NULL,
  pageSize = NULL,
  includeAttributeNames = NULL
)
}
\arguments{
\item{pageNumber}{(optional)  Whole number that specifies which page of results to fetch.}

\item{pageSize}{(optional)  Whole number that specifies the number of results to include per page.}

\item{includeAttributeNames}{(optional)  One or more attribute names}
}
\description{
Lookup Source Attributes
}
\details{
'/content/current/source/SNOMEDCT_US/9468002/attributes'path Retrieves attributes of the SNOMED CT concept and returns a JSON Object classType of Attribute
}
\seealso{
\code{\link[httr]{GET}},\code{\link[httr]{content}}
}
