% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umls_api_lookup_functions.R
\name{lookup_source_descendants}
\alias{lookup_source_descendants}
\title{Lookup Source Descendants}
\usage{
lookup_source_descendants(pageNumber = NULL, pageSize = NULL)
}
\arguments{
\item{pageNumber}{(optional)  Whole number that specifies which page of results to fetch.}

\item{pageSize}{(optional)  Whole number that specifies the number of results to include per page.}
}
\description{
Lookup Source Descendants
}
\details{
'/content/current/source/SNOMEDCT_US/9468002/ancestors'path Retrieves ancestors of a source-asserted identifier and returns a JSON Object classType of SourceAtomCluster*
'/content/current/source/SNOMEDCT_US/9468002/descendants'path Retrieves descendants of a source-asserted identifier and returns a JSON Object classType of SourceAtomCluster*
}
\seealso{
\code{\link[httr]{GET}},\code{\link[httr]{content}}
}
