% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umls_api_lookup_functions.R
\name{lookup_source_parents}
\alias{lookup_source_parents}
\title{Lookup Source Parents}
\usage{
lookup_source_parents(sourceId, sab, pageNumber = NULL, pageSize = NULL)
}
\arguments{
\item{pageNumber}{(optional)  Whole number that specifies which page of results to fetch.}

\item{pageSize}{(optional)  Whole number that specifies the number of results to include per page.}
}
\description{
Lookup Source Parents
}
\details{
'/content/current/source/SNOMEDCT_US/9468002/parents'path Retrieves parents a source-asserted identifier and returns a JSON Object classType of SourceAtomCluster*
'/content/current/source/SNOMEDCT_US/9468002/children'path Retrieves children of a source-asserted identifier and returns a JSON Object classType of SourceAtomCluster*
}
\seealso{
\code{\link[httr]{GET}},\code{\link[httr]{content}}
}
