% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umls_api_lookup_functions.R
\name{lookup_source_relations}
\alias{lookup_source_relations}
\title{Lookup Source Relations}
\usage{
lookup_source_relations(
  sourceId,
  vocabulary,
  includeRelationLabels = NULL,
  includeAdditionalRelationLabels = NULL,
  includeObsolete = NULL,
  includeSuppressible = NULL,
  pageNumber = NULL,
  pageSize = NULL
)
}
\arguments{
\item{includeRelationLabels}{(optional)  One or more relation labels}

\item{includeAdditionalRelationLabels}{(optional)  One or more relation attribute}

\item{includeObsolete}{(optional)  Include content that is obsolete according to the content provider or NLM.}

\item{includeSuppressible}{(optional)  Include content that is suppressible according to NLM Editors.}

\item{pageNumber}{(optional)  Whole number that specifies which page of results to fetch.}

\item{pageSize}{(optional)  Whole number that specifies the number of results to include per page.}
}
\description{
Lookup Source Relations
}
\details{
'/content/current/source/LNC/44255-8/relations'path Retrieves relationships of LOINC code 44255-8 and returns a JSON Object classType of AtomClusterRelation
}
\seealso{
\code{\link[httr]{GET}},\code{\link[httr]{content}}
}
